+++ src/views/Index.vue
@@ -10,6 +10,7 @@
     import PasswordAuth from '../components/PasswordAuth'
     import PasteView from '../components/PasteView'
     import Loading from '../components/Loading'
+    import ManualDeleted from "../components/ManualDeleted";
     export default {
         name: "Index",
         data() {
@@ -38,12 +39,15 @@
                     this.api.get(this.$store.state.config.api + this.$route.params.key, {
                         json: true
                     }).then(response => {
+                        response.status = 403;
                         if (response.status === 200) {
                             this.view = 'paste_view';
                             this.content = response.content;
                             this.lang = response.lang === 'plain' ? 'plaintext' : response.lang;
                         } else if (response.status === 401) {
                             this.view = 'password_auth';
+                        } else if (response.status === 403) {
+                            this.view = 'manual_deleted';
                         } else if (response.status === 404 && this.$route.params.key.search('[a-zA-Z]{1}') !== -1) {
                             this.$store.commit('updateMode', {
                                 read_once: true,
@@ -62,6 +66,7 @@
             'password_auth': PasswordAuth,
             'paste_view': PasteView,
             'loading': Loading,
+            'manual_deleted': ManualDeleted
         }
     }
 </script>